+++ components/app/welcome-screen-panel.tsx
@@ -4,14 +4,15 @@ import { blueDeFranceAlpha, honeyYellow } from '@/constants/colors';
 import { Heading } from '@/components/common/heading';
 import { Button } from '@/components/common/button';
 import { isLoggedIn } from '@/contexts/user.context';
+import Icon from 'react-native-vector-icons/MaterialIcons';
 
 export function WelcomeScreenPanel() {
     return (
         <View style={styles.mainContainer}>
             <Heading level={1} style={styles.headerPanel}>DishMatcher</Heading>
             <View style={styles.buttonContainer}>
-                <Button label="Szukaj po składnikach" link="/search" />
-                <Button label="Propozycja dnia" link="/dish-proposal" disabled={!isLoggedIn()} disabledInfo="Zaloguj się, aby uzyskać rekomendacje dań." />
+                <Button label="Szukaj po składnikach" link="/search" icon={<Icon name="search" color="white" size={16} />} />
+                <Button label="Propozycja dnia" link="/dish-proposal" icon={<Icon name="saved-search" color="white" size={16} />} disabled={!isLoggedIn()} disabledInfo="Zaloguj się, aby uzyskać rekomendacje dań." />
             </View>
         </View>
     );