+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.17.0] - 2024-10-24
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): Black-transparent overlay at landing page.
+- [Marek Kurańda](https://github.com/mjkuranda): `react-native-vector-icons` library.
+- [Marek Kurańda](https://github.com/mjkuranda): More icons.
+- [Marek Kurańda](https://github.com/mjkuranda): Sending ingredients while searching for feeding proposal system.
+
+### Removed
+- [Marek Kurańda](https://github.com/mjkuranda): `Icon` component.
+
 ## [1.16.0] - 2024-10-23
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Smooth screen scrolling at landing page.